Replace PNG logos with SVG
