+++ index.html
@@ -2,12 +2,12 @@
 <html lang="es">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" href="/favicon.ico" />
+    <link rel="icon" type="image/svg+xml" href="/logo.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <meta name="theme-color" content="#0C0D13" />
     <meta name="description" content="Weavion - Desarrollo web y diseño de experiencias digitales" />
-    <link rel="apple-touch-icon" href="/logo192.png" />
-    <link rel="manifest" href="/manifest.json" />
+    <link rel="apple-touch-icon" href="/logo.svg" />
+    <link rel="manifest" href="./manifest.json" />
     <title>Weavion</title>
 
       <!-- Fuente Ars Nova local -->
@@ -15,7 +15,7 @@
       <!-- Preload Ars Nova font files -->
     <link
       rel="preload"
-      href="/fonts/ArsNova-Regular.woff2"
+      href="./fonts/ArsNova-Regular.woff2"
       as="font"
       type="font/woff2"
       crossorigin
@@ -23,7 +23,7 @@
 
     <link
       rel="preload"
-      href="/fonts/ArsNova-Bold.woff2"
+      href="./fonts/ArsNova-Bold.woff2"
       as="font"
       type="font/woff2"
       crossorigin
@@ -32,14 +32,14 @@
     <!-- Preload Argent Pixel CF font files -->
     <link
       rel="preload"
-      href="/fonts/ArgentPixelCF-Regular.woff2"
+      href="./fonts/ArgentPixelCF-Regular.woff2"
       as="font"
       type="font/woff2"
       crossorigin
     />
     <link
       rel="preload"
-      href="/fonts/ArgentPixelCF-Italic.woff2"
+      href="./fonts/ArgentPixelCF-Italic.woff2"
       as="font"
       type="font/woff2"
       crossorigin
